upwork-job-monitor/requirements.txt
requests==2.31.0
schedule
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1